message sending part
